- Fix wrong use of namespace package's __init__ ( cannot have code in root __init__ https://www.python.org/dev/peps/pep-0420/#rationale)
tendrl-bug-id: Tendrl/specifications#37
tendrl-bug-id: Tendrl/specifications#54
tendrl-bug-id: Tendrl/specifications#32
